requirements.txt
anthropic==0.67.0
beautifulsoup4==4.13.5
bert_score==0.3.13
ipython==8.12.3
matplotlib==3.10.6
nltk==3.9.1
numpy==2.3.3
openai==1.107.2
pandas==2.3.2
plotnine==0.15.0
protobuf==6.32.1
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.7.2
seaborn==0.13.2
